scheduler_ly_h_begin
