+++ coroutine-lib-ly/fiber_lib/1thread/thread_ly.h
@@ -78,20 +78,6 @@ private:
     Semaphore m_semaphore;
 };
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
 
 #endif
